Fix float being passed instead of int
I'm not sure why this started happening lately. Maybe it's because I
started looking into Qt6 development, so my global environment might
have changed and broken things. Or the recent switch from python 3.9 to
3.10 might have caused it. Or the Qt5 to Qt6 transition. In any case,
this function was complaining that it got passed a `float` when it
expected an `int`, so the natural thing to do is to cast.